% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fixed_effect_a_criterion.R
\name{a_criterion}
\alias{a_criterion}
\title{Description of a_criterion_f function}
\usage{
a_criterion(Xt, D = diag(1, nrow = nrow(Xt), ncol = nrow(Xt)), P)
}
\arguments{
\item{Xt}{is the design matrix of treatments}

\item{D}{is the design matrix for the estimation, initial as a identity matrix.}
}
\value{
Returning the value of the A-criteria for this design.
}
\description{
This function calculate A-criteria for a specific design.
}
