% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pairwise_permutations.R
\name{generate_pairwise_permutations}
\alias{generate_pairwise_permutations}
\title{Description of a_criterion_f function}
\usage{
generate_pairwise_permutations(X)
}
\arguments{
\item{X}{is the matrix we need all its pairwise permutations}
}
\value{
A list of matrices that contains all pairwise permutations of X matrix
}
\description{
A function to generate all pairwise permutations of rows for a matrix. If two rows are same,
we skip it the permutation doesn't change the initial matrix
}
