% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_matrix_in_list.R
\name{is_matrix_in_list}
\alias{is_matrix_in_list}
\title{Description of is_matrix_in_list function}
\usage{
is_matrix_in_list(matrix, list_of_matrices)
}
\arguments{
\item{matrix}{is the matrix we want to check}

\item{list_of_matrices}{is the list of matrices}
}
\value{
A bool value. Returning TRUE if matrix is in the list_of_matrices, False if it is not
}
\description{
A function to check if a matrix is in a matrix list.
}
