% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_matrix.R
\name{P_matrix}
\alias{P_matrix}
\title{Description of P_matrix function}
\usage{
P_matrix(Z, R, G)
}
\arguments{
\item{Z}{is the design matrix of the random part of the model, such as row and column effects.}

\item{R}{is the variance matrix of the model}

\item{G}{is the variance matrix for random part}
}
\value{
Returning a P matrix for the model.
}
\description{
To calculate P matrix for a model
}
