% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fixed_effect_a_criterion.R
\name{P_matrix}
\alias{P_matrix}
\title{Description of a_criterion_f function}
\usage{
P_matrix(Z, R, G)
}
\arguments{
\item{Xt}{is the design matrix of treatments}

\item{D}{is the design matrix for the estimation, initial as a identity matrix.}
}
\value{
Returning the value of the A-criteria for this design.
}
\description{
This function calculate A-criteria for a specific design.
}
